More accurate caret placement and ideavim support (#25)
* More accurate caret movement

When I made block mode work, I saw that the caret is barely missing drawing in the right place.

It was largely reproducible if you move left then right and then try moving right then left. The caret would always settle in a barely (~1-2px) different position depending on which direction the animation was coming from.

The issue was 3 fold:
- We're using `visualPositionToXY` which is less accurate since it spits out `int`s (inside a `Point`) and then casting to double, losing more precision
- We're using `toInt()` directly which always floors, so sometimes we'd be just ever so slightly off to the left
- We're stopping animation before reaching the final spot (`0.01` instead of `0`), so sometimes the animation stops just a bit short of reaching the correct final resting place.

Using `visualPositionToPoint2D` to get the accurate visual position, `round` to be able to round up if needed, and waiting until `dx=0` to finish the animation resolves the inaccuracies.

* Use editor caret attributes to determine caret shape

Instead of always forcing one shape or another, we can just check what the IDE has as the caret shape, and make a shape like that.

This makes it so that the plugin is compatible with e.g. Ideavim which likes to switch back and forth between BLOCK and LINE/DEFAULT.

With this change, we can also get rid of the `CaretStyle` settings, since we just follow `CaretVisualAttributes.Shape`.

It seems `SmoothCaretSettings.CaretStyle.BLOCK` is exactly the same as `SmoothCaretSettings.CaretStyle.LINE`, so I instead wrote a change that will try to measure the next character and draw a block caret of that size (or fall back to using 'm' to measure).

* Full width underscore

The method here to draw an underscore only drew a 2x bar width underscore. Update it to use the full width of the next character.

This brings it a bit closer to the standard underscore cursor:

<img width="28" height="29" alt="image" src="https://github.com/user-attachments/assets/f7e52f02-b0ed-4dfc-a3e4-cbd9ae7905fc" />

<img width="22" height="22" alt="image" src="https://github.com/user-attachments/assets/1c0f5005-d69a-458a-9b51-52e9dc1301c1" />